YARN-679 give all tests a TestService to identify their role and aid bulk mvn test runs
